requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.7
daphne==4.1.2
defusedxml==0.8.0rc2
Django==5.0.4
django-allauth==0.63.3
django-cors-headers==4.3.1
django-environ==0.11.2
django-phonenumber-field==7.3.0
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
oauthlib==3.2.2
phonenumbers==8.13.37
pillow==10.3.0
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.1.0
setuptools==70.3.0
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.4.4
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
zope.interface==6.4.post2
